moving rolesEq to a separate file due to build error
