+++ package.json
@@ -16,34 +16,33 @@
     "logs:liquibase": "docker compose --env-file .env.local logs -f liquibase"
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-navigation-menu": "^1.2.14",
     "@radix-ui/themes": "^3.2.1",
     "@tanstack/react-query": "^5.85.6",
     "next": "16.0.0",
     "next-themes": "^0.4.6",
     "pg": "^8.16.3",
     "react": "19.2.0",
-    "react-dom": "19.2.0"
+    "react-dom": "19.2.0",
+    "server-only": "^0.0.1"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@radix-ui/react-icons": "^1.3.2",
     "@tailwindcss/postcss": "^4",
     "@tanstack/eslint-plugin-query": "^5.83.1",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^15.0.0",
+    "@types/jest": "^29.5.11",
     "@types/node": "^20",
     "@types/pg": "^8.15.5",
     "@types/react": "19.2.2",
     "@types/react-dom": "19.2.2",
     "better-auth": "^1.3.30",
-    "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/react": "^15.0.0",
-    "@types/jest": "^29.5.11",
     "eslint": "^9",
     "eslint-config-next": "16.0.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "identity-obj-proxy": "^3.0.0",
     "next-intl": "^4.5.2",
     "prettier": "^3.6.2",
     "tailwindcss": "^4",
